seperate v_delta_precalc out
